Refactor code into separate files, target c# preview lang version
