+++ ChatApp.Server/DbContexts/ApplicationDbContext.cs
@@ -1,0 +1,13 @@
+using ChatApp.Server.Models;
+using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore;
+
+namespace ChatApp.Server.DbContexts
+{
+    public class ApplicationDbContext : IdentityDbContext<ApplicationUser>
+    {
+        public ApplicationDbContext(DbContextOptions options) : base(options)
+        {
+        }
+    }
+}
