+++ ChatApp.Server/Models/DTO/UserDTO.cs
@@ -1,0 +1,8 @@
+namespace ChatApp.Server.Models.DTO
+{
+    public class UserDTO
+    {
+        public string Id { get; set; }
+        public string UserName { get; set; }
+    }
+}
